exit_status

+++ src/heredoc_2.c
@@ -1,17 +1,33 @@
 #include "../includes/minishell.h"
 
-static void	set_pipe_fd_and_write(int heredoc_pipe_fd[2], int tmp_fd, \
+static int	set_pipe_fd_and_write(int heredoc_pipe_fd[2], int tmp_fd, \
 	char *cat_line, char *line)
 {
-	close(heredoc_pipe_fd[0]);
-	dup2(heredoc_pipe_fd[1], tmp_fd);
-	close(heredoc_pipe_fd[1]);
+	if (close(heredoc_pipe_fd[0]) == -1)
+	{
+		perror(SHELLNAME);
+		return (ERROR);
+	}
+	if (dup2(heredoc_pipe_fd[1], tmp_fd) == -1)
+	{
+		perror(SHELLNAME);
+		return (ERROR);
+	}
+	if (close(heredoc_pipe_fd[1]) == -1)
+	{
+		perror(SHELLNAME);
+		return (ERROR);
+	}
 	write(tmp_fd, cat_line, ft_strlen(cat_line));
-	close(tmp_fd);
+	if (close(tmp_fd) == -1)
+	{
+		perror(SHELLNAME);
+		return (ERROR);
+	}
 	if (unlink("tmp.txt") == -1)
 	{
 		perror(SHELLNAME);
-		exit(ERROR);
+		return (ERROR);
 	}
 	free(line);
 	free(cat_line);
@@ -58,7 +74,7 @@ static int	open_fd_and_calloc(int *tmp_fd, char **cat_line, int *flag)
 	return (SUCCESS);
 }
 
-int	heredoc_parent_process(t_info *info, int heredoc_pipe_fd[2], \
+int	heredoc_child_process(t_info *info, int heredoc_pipe_fd[2], \
 	int continue_flag)
 {
 	int		flag;
@@ -77,8 +93,11 @@ int	heredoc_parent_process(t_info *info, int heredoc_pipe_fd[2], \
 		if (ft_strcmp(line, info->sentence_lst->redirect_lst->str) == 0)
 		{
 			if (continue_flag == 0)
-				set_pipe_fd_and_write(heredoc_pipe_fd, tmp_fd, \
-					cat_line, line);
+			{
+				if (set_pipe_fd_and_write(heredoc_pipe_fd, tmp_fd, \
+					cat_line, line) == ERROR)
+					return (ERROR);
+			}
 			return (SUCCESS);
 		}
 		free(line);
